Node demo consolidation, web demo still works with 1.6.1
