+++ browser/basic-react-app/package.json
@@ -7,7 +7,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "@trinsic/okapi": "^1.0.0-preview.27",
+    "@trinsic/okapi-web": "^1.6.1-rc1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
