+++ browser/basic-webpack/package.json
@@ -11,10 +11,9 @@
   "author": "Tomislav Markovski",
   "license": "ISC",
   "dependencies": {
-    "@trinsic/okapi": "^1.0.0-preview.27"
+    "@trinsic/okapi-web": "^1.6.1-rc1"
   },
   "devDependencies": {
-    "@types/google-protobuf": "^3.15.3",
     "http-server": "^0.12.3",
     "webpack": "^5.38.1",
     "webpack-cli": "^4.7.1"
