--- node/basic/README.md
@@ -1,8 +1,0 @@
-## `okapi` for Node
-
-This sample shows how `okapi` can be used in Node.
-
-To run the sample, build and run using
-
-- `npm i`
-- `node index.js`
